Split webhooks related flows in separated files and folders
