+++ backend/infrahub/webhook/tasks/__init__.py
@@ -1,0 +1,8 @@
+from .configure import configure_webhook
+from .process import convert_node_to_webhook, webhook_process
+
+__all__ = [
+    "configure_webhook",
+    "convert_node_to_webhook",
+    "webhook_process",
+]
